chore: protocol 소스파일 분리
