+++ HeartHub/CommunityScene/View/Protocol/CommunityCellProfileViewDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  CommunityCellProfileViewDelegate.swift
+//  HeartHub
+//
+//  Created by 이태영 on 2023/07/19.
+//
+
+import Foundation
+
+protocol CommunityCellProfileViewDelegate: AnyObject {
+    func didTapUserProfile()
+    func didTapPostOption()
+}
